requirements.txt
Click==7.0
hurry.filesize==0.9
